+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,6 @@
             android:networkSecurityConfig="@xml/network_security_config"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
-            android:name=".course.database.RoomApplication"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
@@ -22,12 +21,11 @@
                 android:theme="@style/AppTheme.Launcher">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
-
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".course.views.CreateNewUserView"></activity>
-        <activity android:name=".course.views.PokemonView"></activity>
+        <activity android:name=".course.views.CreateNewUserView" />
+        <activity android:name=".course.views.PokemonView" />
     </application>
 
 </manifest>
